DMIDATA Plugin
Initial checkin for the dmidata plugin

This plugin is expected to read the Desktop Management Interface entries
located under /sys/* either directly or using the help of hwloc and push
the data to the aggregator.

The primary use of this component is for inventory purposes. It is
expected to read the list of components, and their unique identifiers
and log them into the inventory database. It will also keep track of the
hot-swappable h/w and note any changes and log them into the databse

Another minor change involves adding ignores w.r.t. temporary files
generated by vim and slickedit